:recycle: Split player component
