+++ Assets/Resources/NeoC/Model.prefab
@@ -216,6 +216,7 @@ GameObject:
   - component: {fileID: 4000011149700772}
   - component: {fileID: 95000013542507340}
   - component: {fileID: 114366912592162476}
+  - component: {fileID: 114196709688271170}
   - component: {fileID: 54606530159798840}
   - component: {fileID: 114337247030454140}
   m_Layer: 0
@@ -1049,6 +1050,19 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1a2796ff9e6924e0c851084d14e507fa, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &114196709688271170
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1000012414845946}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bb615df0c1326444a963e5448bc65834, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _animator: {fileID: 95000013542507340}
+  speed: 0.1
 --- !u!114 &114337247030454140
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1062,6 +1076,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _components:
   - {fileID: 114366912592162476}
+  - {fileID: 114196709688271170}
   _identifier: 
   _context: {fileID: 0}
   _bindType: 0
@@ -1077,7 +1092,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _animator: {fileID: 95000013542507340}
-  speed: 0.1
 --- !u!135 &135375320855292552
 SphereCollider:
   m_ObjectHideFlags: 1
