+++ Assets/Scripts/Game/Attacker.cs
@@ -4,10 +4,9 @@ using UniRx.Triggers;
 
 namespace NeoC.Game
 {
-    public class PlayerController : ObservableTriggerBase
+    public class Attacker : ObservableTriggerBase
     {
         [SerializeField] private Animator _animator;
-        [SerializeField] private float speed;
 
         private Subject<Saboten> onAttack;
 
@@ -21,19 +20,6 @@ namespace NeoC.Game
 
             OnAttackAsObservable()
                 .Subscribe(_ => TriggerAnimator("Attack"));
-        }
-
-        public void Move(Vector2 move)
-        {
-            move *= speed;
-            Move(move.X0Y());
-        }
-
-        public void Move(Vector3 move)
-        {
-            transform.position += move;
-            transform.LookToward(-move);
-            _animator.SetFloat("velocity", move.sqrMagnitude);
         }
 
         private void TriggerAnimator(string name)
