+++ Assets/Scripts/Game/PlayerMover.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bb615df0c1326444a963e5448bc65834
+timeCreated: 1482845140
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
